requirements.txt
asn1crypto==0.24.0
astroid==2.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3
Django==2.0.6
django-cors-middleware==1.2.0
django-extensions==1.7.1
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pycparser==2.18
PyJWT==1.6.4
pylint==2.0.0
pytz==2018.5
requests==2.19.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.23
wrapt==1.10.11
